Rewrite clr-power-tweaks to use tmpfiles.d.
Most of the tweaks are simply writes to files in /proc & /sys. Use
tmpfiles.d(5) snippet for use with the "w" stanza (write to file, if
exists with glob patterns). This significantly reduces the codebase,
drops --vm service file & mode of operation.

Remaining functionality in the clr_power binary is more dynamic stuff:
pci & usb power management, disabling of wake-on-lan.