Makefile.am: better dependency tracking, cleanup
Ensure that dependency tracking is enabled for all headers and config
files. Add them to SOURCES for tracking. Because we are doing this,
autotools documentation recommends we not add headers to a 'noinst'
target any longer.

Move the README.md content into the README file, and add a symlink to
README.md for GitHub's sake. As a side-effect, README (and all its
content) will now be included in the 'make dist' target.

Remove COPYING from EXTRA_DIST, as that is one of the files
automatically picked up by autotools.

Signed-off-by: Joe Konno <joe.konno@intel.com>