Convert generated list to inline data.
This is needed to prepare for some sort of user-local override
mechanic - we need to be able to go through the list and avoid
writing user-specified entries. The easiest way is to have the
data available as a simple list we can iterate through in memory
so that we can skip items.

This should be easy enough to edit & compile as before.