Add local overrides/excludes.
Introduces "/etc/clr-power-tweaks.conf" which allows to
prevent default tweaks from being written and/or add new
custom tweaks.

The custom tweaks are applied before the builtin ones. Either
nothing is written to a match or file, and the file is added to
the exclude list, or the value listed is written to the globbed
files and it is added to the exclude list.