Add desktop vs. server specific tweaking.
We introduce a simple flag to tell whether files are specific
to server, or desktop, or neutral (assumed server).

The default is to tune for servers. If we don't know what a system
is, we'll treat it like a server for now.

You can tell it to tune for desktop by adding `desktop` as a startup
flag on the command line. It will autodetect Intel desktop client
systems by model name, and treats Core(TM), Celeron and Pentium
named CPU's as desktop systems. On systems that are detected to be
a desktop system, you can pass `server` instead.

Overrides continue to work as normal.

As a concept illustration, we set the NVME queue scheduler algorithm
to BFQ low latency for desktops. All other settings remain the same.