Remove somaxconn override.
Previously, the default was 128, which we conservatively increased
to a higher value. 5.4.x kernels however have a value of 4096 by
default, meaning we *decrease* the amount, unintentionally.

It's possible that this may affect something negatively, but, if
that is found to be, we can address it when needed. This should
largely be a positive change in most cases.